Add files via upload
The code runs under Windows XP or previous versions. It draws the paths of the polluted injected in a generic grid node on the computer screen during the simulation. The output gives the file "RIESCE" with a numerical description of the flow path followed by any single particle, and the file "TEMPO". The latter shows the time required from any specific particle to get a specific grid node. The histogram of these times provides the expected breakthrough curve of the pollutant in the selected node. Details about the numerical technique (particle tracking) can be found in Masciopinto et al. (2010), doi:10.1016/j.jconhyd.2010.07.003.